feat: refactor booking module and enhance booking functionality with user role checks

+++ src/modules/user/user.service.ts
@@ -1,11 +1,19 @@
+import { JwtPayload } from "jsonwebtoken";
 import { pool } from "../../config/db";
 
 const getUsers = async () => {
   return await pool.query("SELECT id, name, email, phone, role FROM users");
 };
 
-const updateUser = async (userId: string, payload: Record<string, unknown>) => {
+const updateUser = async (
+  userId: string,
+  payload: Record<string, unknown>,
+  user: JwtPayload
+) => {
   const { name, email, phone, role } = payload;
+
+  if (user.role === "customer" && userId != user.id)
+    throw new Error("Forbidden access");
 
   return await pool.query(
     `
@@ -18,6 +26,15 @@ const updateUser = async (userId: string, payload: Record<string, unknown>) => {
 };
 
 const deleteUser = async (userId: string) => {
+  const bookingResult = await pool.query(
+    "SELECT status FROM bookings WHERE customer_id = $1",
+    [userId]
+  );
+
+  if (bookingResult.rows[0].status === "active") {
+    throw new Error("Cannot delete user with active booking");
+  }
+
   return await pool.query("DELETE FROM users WHERE id = $1", [userId]);
 };
